requirements.txt
pytest==7.4.2
allure-pytest==2.13.2
pytest-cov==4.1.0
